requirements.txt
# Web框架
flask>=2.0.0
werkzeug>=2.0.0

# Word处理
python-docx>=0.8.11

# Excel处理
pandas>=2.0.0
openpyxl>=3.0.0

# LLM服务
requests>=2.28.0

# 环境变量
python-dotenv>=1.0.0
